+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 - Binance parser: added config to split BNB quantity across multiple tokens.
 - Conversion tool: added paser for StakeTax.
 - Nexo parser: added new export format. ([#286](https://github.com/BittyTax/BittyTax/issues/286))
+- Etherscan parser: added new export formats. ([#290](https://github.com/BittyTax/BittyTax/issues/290))
 ### Changed
 - Binance parser: performance improvements for large data sets.
 - Coinbase Pro parser: performance improvements for large data sets.